finish semctl.setval. pass compilation.

+++ kernel/src/syscall/mod.rs
@@ -264,7 +264,7 @@ impl Syscall<'_> {
             // sem
             SYS_SEMGET => self.sys_semget(args[0], args[1], args[2]),
             SYS_SEMOP => self.sys_semop(args[0], args[1] as *const SemBuf, args[2]),
-            SYS_SEMCTL => self.sys_semctl(args[0], args[1], args[2], args[3]/* as SemctlUnion*/),
+            SYS_SEMCTL => self.sys_semctl(args[0], args[1], args[2], args[3] as isize/* as SemctlUnion*/),
 
             // system
             SYS_GETPID => self.sys_getpid(),
